requirements.txt
Flask==2.2.2
Flask_MySQLdb==1.0.1
requests==2.28.1
PyYAML~=6.0
